pyproject.toml
[tool.poetry]
name = "local-test"
version = "0.1.0"
description = ""
authors = ["Matt Barkway <matt.barkway@ovo.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "<3.14,>=3.8,"
pyyaml = "^6.0.2"
pydantic = "^2.9.1"
colorama = "^0.4.6"
typer = "^0.12.5"
pyinstaller = "^6.10.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
